replace int with long to enlarge taxonomy size

+++ tiwHTMLversion/src/main/java/it/polimi/tiw/beams/Category.java
@@ -1,12 +1,12 @@
 package it.polimi.tiw.beams;
 
 public class Category implements Comparable<Category>{
-    private int code;
+    private long code;
     private String name;
     private int level;
     private int padding;
     private boolean selected;
-    public Category(int code, String name){
+    public Category(long code, String name){
         this.code = code;
         this.name = name;
         selected = false;
@@ -27,7 +27,7 @@ public class Category implements Comparable<Category>{
     /*
      *       get methods
      */
-    public int getCode() {
+    public long getCode() {
         return code;
     }
 
@@ -87,6 +87,6 @@ public class Category implements Comparable<Category>{
      */
     @Override
     public int compareTo(Category o) {
-        return (Integer.toString(this.getCode()).compareTo(Integer.toString(o.getCode())));
+        return (Long.toString(this.getCode()).compareTo(Long.toString(o.getCode())));
     }
 }
